+++ CHANGELOG.md
@@ -6,7 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 
 ## [TODO]
 
-## [Unreleased]
+## [1.17.0] - 2025-02-23
 
 ### Added
 - Add `returns` macro for context output validation after task execution